[wip] backend: partial structure refactor
