+++ moly-backend/src/controllers/mod.rs
@@ -1,0 +1,4 @@
+pub mod download;
+pub mod download_files;
+pub mod model_cards;
+pub mod models;
