+++ moly-backend/src/controllers/model_cards.rs
@@ -2,11 +2,11 @@ use std::collections::HashMap;
 use std::io::{self, Write};
 use std::path::{Path, PathBuf};
 use std::str;
-use std::sync::Arc;
 
-use chrono::{DateTime, Utc};
 use git2::{FetchOptions, ProxyOptions, Repository};
 use serde::{Deserialize, Serialize};
+
+use crate::models::ModelCard;
 
 fn do_fetch<'a>(
     repo: &'a git2::Repository,
@@ -362,142 +362,6 @@ impl ModelIndex {
         model_card.download_count = self.download_count;
 
         Ok(model_card)
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq, Deserialize, Serialize)]
-pub struct Author {
-    pub name: String,
-    pub url: String,
-    pub description: String,
-}
-
-impl Into<moly_protocol::data::Author> for Author {
-    fn into(self) -> moly_protocol::data::Author {
-        moly_protocol::data::Author {
-            name: self.name,
-            url: self.url,
-            description: self.description,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct ModelCard {
-    pub id: String,
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub summary: String,
-    #[serde(default)]
-    pub size: String,
-    #[serde(default)]
-    pub requires: String,
-    #[serde(default)]
-    pub architecture: String,
-    pub released_at: DateTime<Utc>,
-    #[serde(default)]
-    pub files: Vec<RemoteFile>,
-    pub prompt_template: String,
-    pub reverse_prompt: String,
-    pub context_size: u64,
-    pub author: Author,
-    #[serde(default)]
-    pub like_count: u32,
-    #[serde(default)]
-    pub download_count: u32,
-    #[serde(default)]
-    pub metrics: Option<HashMap<String, f32>>,
-}
-
-#[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize)]
-pub struct RemoteFile {
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub size: String,
-    #[serde(default)]
-    pub quantization: String,
-    #[serde(default)]
-    pub tags: Vec<String>,
-    #[serde(default)]
-    pub sha256: Option<String>,
-    #[serde(default)]
-    pub download: HashMap<String, String>,
-}
-
-impl ModelCard {
-    pub fn to_model(
-        remote_models: &[Self],
-        conn: &rusqlite::Connection,
-    ) -> rusqlite::Result<Vec<moly_protocol::data::Model>> {
-        let model_ids = remote_models
-            .iter()
-            .map(|m| m.id.clone())
-            .collect::<Vec<_>>();
-        let files = super::download_files::DownloadedFile::get_by_models(conn, &model_ids)?;
-
-        fn to_file(
-            model_id: &str,
-            remote_files: &[RemoteFile],
-            save_files: &HashMap<Arc<String>, super::download_files::DownloadedFile>,
-        ) -> rusqlite::Result<Vec<moly_protocol::data::File>> {
-            let mut files = vec![];
-            for remote_f in remote_files {
-                let file_id = format!("{}#{}", model_id, remote_f.name);
-                let downloaded_path = save_files.get(&file_id).map(|file| {
-                    let file_path = Path::new(&file.download_dir)
-                        .join(&file.model_id)
-                        .join(&file.name);
-                    file_path
-                        .to_str()
-                        .map(|s| s.to_string())
-                        .unwrap_or_default()
-                });
-
-                let file = moly_protocol::data::File {
-                    id: file_id,
-                    name: remote_f.name.clone(),
-                    size: remote_f.size.clone(),
-                    quantization: remote_f.quantization.clone(),
-                    downloaded: downloaded_path.is_some(),
-                    downloaded_path,
-                    tags: remote_f.tags.clone(),
-                    featured: false,
-                };
-
-                files.push(file);
-            }
-
-            Ok(files)
-        }
-
-        let mut models = Vec::with_capacity(remote_models.len());
-
-        for remote_m in remote_models {
-            let model = moly_protocol::data::Model {
-                id: remote_m.id.clone(),
-                name: remote_m.name.clone(),
-                summary: remote_m.summary.clone(),
-                size: remote_m.size.clone(),
-                requires: remote_m.requires.clone(),
-                architecture: remote_m.architecture.clone(),
-                released_at: remote_m.released_at.clone(),
-                files: to_file(&remote_m.id, &remote_m.files, &files)?,
-                author: moly_protocol::data::Author {
-                    name: remote_m.author.name.clone(),
-                    url: remote_m.author.url.clone(),
-                    description: remote_m.author.description.clone(),
-                },
-                like_count: remote_m.like_count.clone(),
-                download_count: remote_m.download_count.clone(),
-                metrics: remote_m.metrics.clone().unwrap_or_default(),
-            };
-
-            models.push(model);
-        }
-
-        Ok(models)
     }
 }
